Split v2 domain models by topic
