+++ policyengine_api/data/v2/models/associations.py
@@ -1,0 +1,129 @@
+"""Canonical SQLModel tables linking v2 users to their domain records."""
+
+from datetime import datetime
+from typing import TYPE_CHECKING
+from uuid import UUID
+
+import sqlalchemy as sa
+from sqlmodel import Field, Relationship
+
+from policyengine_api.data.v2.models.base import TimestampedModel
+from policyengine_api.data.v2.models.households import Household
+from policyengine_api.data.v2.models.policies import Policy
+from policyengine_api.data.v2.models.simulations import Simulation
+from policyengine_api.data.v2.models.users import User
+
+if TYPE_CHECKING:
+    from policyengine_api.data.v2.models.reports import Report
+
+
+class UserHouseholdAssociation(TimestampedModel, table=True):
+    __tablename__ = "user_household_associations"
+    __table_args__ = (
+        sa.UniqueConstraint(
+            "user_id",
+            "household_id",
+            name="uq_user_household_associations_user_household",
+        ),
+    )
+
+    user_id: UUID = Field(
+        foreign_key="users.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    household_id: UUID = Field(
+        foreign_key="households.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    country: str = Field(max_length=16)
+    label: str | None = Field(default=None, max_length=255)
+
+    user: User = Relationship(back_populates="household_associations")
+    household: Household = Relationship(back_populates="user_associations")
+
+
+class UserPolicy(TimestampedModel, table=True):
+    __tablename__ = "user_policies"
+    __table_args__ = (
+        sa.UniqueConstraint(
+            "user_id",
+            "policy_id",
+            name="uq_user_policies_user_policy",
+        ),
+    )
+
+    user_id: UUID = Field(
+        foreign_key="users.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    policy_id: UUID = Field(
+        foreign_key="policies.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    country: str = Field(max_length=16)
+    label: str | None = Field(default=None, max_length=255)
+
+    user: User = Relationship(back_populates="policy_associations")
+    policy: Policy = Relationship(back_populates="user_associations")
+
+
+class UserSimulationAssociation(TimestampedModel, table=True):
+    __tablename__ = "user_simulation_associations"
+    __table_args__ = (
+        sa.UniqueConstraint(
+            "user_id",
+            "simulation_id",
+            name="uq_user_simulation_associations_user_simulation",
+        ),
+    )
+
+    user_id: UUID = Field(
+        foreign_key="users.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    simulation_id: UUID = Field(
+        foreign_key="simulations.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    country: str = Field(max_length=16)
+    label: str | None = Field(default=None, max_length=255)
+
+    user: User = Relationship(back_populates="simulation_associations")
+    simulation: Simulation = Relationship(back_populates="user_associations")
+
+
+class UserReportAssociation(TimestampedModel, table=True):
+    __tablename__ = "user_report_associations"
+    __table_args__ = (
+        sa.UniqueConstraint(
+            "user_id",
+            "report_id",
+            name="uq_user_report_associations_user_report",
+        ),
+    )
+
+    user_id: UUID = Field(
+        foreign_key="users.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    report_id: UUID = Field(
+        foreign_key="reports.id",
+        ondelete="CASCADE",
+        index=True,
+    )
+    country: str = Field(max_length=16)
+    label: str | None = Field(default=None, max_length=255)
+    last_run_at: datetime | None = Field(
+        default=None,
+        sa_type=sa.DateTime(timezone=True),
+    )
+
+    user: User = Relationship(back_populates="report_associations")
+    report: "Report" = Relationship(back_populates="user_associations")
